Update Controllers/BeersControllers.cs
By changing the return type to IActionResult, anyone reading the code will know that the action returns a result that can be turned into an HTTP response (what IActionResult is designed to represent!) ActionResult<T>, says that the action will return either a response with a type of T or one of the various HTTP status results (like NotFound etc). This is more specific (better for this case) and works best when I know that the successful response will always contain data of a specific type.

Co-authored-by: So Su <56776579+ssu8@users.noreply.github.com>